\name{makefunction}
\alias{makefunction}
\title{Make a function}
\usage{
  makefunction(myfun)
}
\arguments{
  \item{myfun}{a function}
}
\value{
  the same function
}
\description{
  This does nothing else than verifying that an object is a
  function, and remove any custom environment.
}
\author{
  Jeroen Ooms
}

