\name{testplothandlers}
\alias{testplothandlers}
\title{Test opencpu plotting}
\usage{
  testplothandlers()
}
\value{
  should return NULL and exactly 6 plots.
}
\description{
  The purpose of this function is to test if opencpu is
  correctly capturing different types of plots.
}
\examples{
testplothandlers(); #should generate six plots
# Equivalent to:
# http://beta1.opencpu.org/R/call/opencpu.demo/testplothandlers/save
}
\author{
  Jeroen Ooms
}

